+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - [web] better responsive layout for the filtersbox.
+- [api] rename param approval to approvals and make it comma separated list.
 - [web] reworked the look and feel of changes tables for a better readability
 - [web] set bootstrap table with size=sm for a better look and feel
 - [crawler] github application: use authlib.jose.jwt instead of pyjwt.